% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_set.R
\name{check_arg_set}
\alias{check_arg_set}
\title{Checks an arg_set (skeleton only)}
\usage{
check_arg_set(arg_set, the_seq = NULL, required_fitness = TRUE,
  verbose = FALSE)
}
\arguments{
\item{arg_set}{The arg_set to test}

\item{required_fitness}{Should the required_fitness element be present in the arg_sets?}

\item{verbose}{Prints out result of each check}
}
\description{
Checks an arg_set (skeleton only)
}
