% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsum.R
\name{check_dcollection_df}
\alias{check_dcollection_df}
\title{Checks the result of dcollection_to_df}
\usage{
check_dcollection_df(dcollection_df, dcollection = NULL)
}
\arguments{
\item{dcollection_df}{The result returned by dcollection_to_df.}

\item{dcollection}{The dcollection on which dcollection_to_df was called.}
}
\description{
Checks that the output produced by a call of dcollection_to_df on a
dcollection produced the expected result. Optionally, the dcollection on
which dcollection_to_df was called can be included to check that the values
are present in the result.
}
