% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsum.R
\name{check_dsum}
\alias{check_dsum}
\title{Check a dsum list}
\usage{
check_dsum(dsum, identifiers = FALSE)
}
\arguments{
\item{dsum}{The dsum to check}

\item{identifiers}{If TRUE, then the dsum will be checked to ensure that
the identifier elements are also present.}
}
\value{
A list with TRUE or FALSE indicating whether the related check
passed.
}
\description{
Checks that a dsum list adheres to the specifications
}
