% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\name{check_fitness_evaluator}
\alias{check_fitness_evaluator}
\title{General fitness_evaluator checker}
\usage{
check_fitness_evaluator(fun, args)
}
\arguments{
\item{fun}{The fitness_evaluator function}

\item{args}{A single list of arguments for the fitness_evaluator that are reasonable. This will be used to call the evaluator on dummy data during the checks.}
}
\description{
Performs basic sanity checks on a fitness_evaluator and its candidate argument set.
}
\details{
Since fitness_evaluators are custom functions that are only used in deeply nested code
and called with mechanisms like do.call, it makes debugging them really
hard. This function performs a number of checks to ensure that the basic
errors are flagged before the main loop is initiated. This function is also used extensively in the testing of the fitness_evaluators.
}
