% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\name{check_fitness_evaluator_result}
\alias{check_fitness_evaluator_result}
\title{General checks for the result of a fitness evaluator}
\usage{
check_fitness_evaluator_result(fitness_evaluation, input_args)
}
\arguments{
\item{fitness_evaluation}{The result returned by a fitness_evaluator}

\item{input_args}{The arguments that were used to all the fitness_evaluator}
}
\value{
A list with a series of boolean elements indicating whether a specific check passed or failed.
}
\description{
General checks that should be satisfied by all fitness_evaluators. No checks specific to the mechanism by which the fitness_score was computed.
}
