% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_processing_metrics.R
\name{check_fitness_processing_metrics}
\alias{check_fitness_processing_metrics}
\title{Checks that a list tracking the fitness_processing_metrics is valid}
\usage{
check_fitness_processing_metrics(fitness_processing_metrics,
  verbose = FALSE)
}
\arguments{
\item{fitness_processing_metrics}{The fitness_processing_metrics list to check}
}
\description{
Checks that a list tracking the fitness_processing_metrics is valid
}
