% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genealogy.R
\name{check_genealogy}
\alias{check_genealogy}
\title{Check genealogy}
\usage{
check_genealogy(genealogy)
}
\arguments{
\item{genealogy}{The genealogy data.frame that must be checked.}
}
\value{
A list with TRUE or FALSE indicating whether the related check passed.
}
\description{
Checks a genealogy for correctness by ensuring that all the columns are
present and that they do not violate any of the structural requirements
(like referencing a parent that does not exist)
}
