% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genealogy.R
\name{check_genealogy_fitness_score}
\alias{check_genealogy_fitness_score}
\title{Check the fitness_score column in a genealogy}
\usage{
check_genealogy_fitness_score(genealogy, result = list())
}
\arguments{
\item{genealogy}{The genealogy to check.}

\item{result}{The list to which the result will be added and from which
previous result will be drawn to check the prerequisites.}
}
\value{
A list with TRUE or FALSE indicating whether the related check
passed.
}
\description{
Checks that fitness_score column contains only valid fitness_scores.
}
