% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genealogy.R
\name{check_genealogy_gen_num}
\alias{check_genealogy_gen_num}
\title{Check the gen_num column in a genealogy}
\usage{
check_genealogy_gen_num(genealogy, results = list())
}
\arguments{
\item{genealogy}{The genealogy to check.}

\item{results}{The list to which the results will be added and from which previous results will be drawn to check the prerequisites.}
}
\description{
Checks that gen_num is not missing, is a natural number and contains all numbers between zero and the max gen_num.
}
