% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genealogy.R
\name{check_genealogy_id}
\alias{check_genealogy_id}
\title{Check the id column in a genealogy}
\usage{
check_genealogy_id(genealogy, results = list())
}
\arguments{
\item{genealogy}{The genealogy to check.}

\item{results}{The list to which the results will be added and from which previous results will be drawn to check the prerequisites.}
}
\description{
Checks that gen_num is not missing, is greater than zero and is an integer.
}
