% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genealogy.R
\name{check_genealogy_n_mut}
\alias{check_genealogy_n_mut}
\title{Check the n_mut column in a genealogy}
\usage{
check_genealogy_n_mut(genealogy, results = list())
}
\arguments{
\item{genealogy}{The genealogy to check.}

\item{results}{The list to which the results will be added and from which
previous results will be drawn to check the prerequisites.}
}
\value{
A list with TRUE or FALSE indicating whether the related check
passed.
}
\description{
Checks that n_mut contains only valid nucleic letters.
}
