% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genealogy.R
\name{check_genealogy_structure}
\alias{check_genealogy_structure}
\title{Check the column names in a genealogy}
\usage{
check_genealogy_structure(results, genealogy)
}
\arguments{
\item{results}{The list to which the results will be added and from which previous results will be drawn to check the prerequisites.}

\item{genealogy}{The genealogy to check.}
}
\description{
Checks that all the required column (and none but these columns) are present in a genealogy.
}
