% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_proc_many_pops.R
\name{check_many_pops}
\alias{check_many_pops}
\title{Checks the results produced by sim_proc_many_pops}
\usage{
check_many_pops(many_pops, verbose = FALSE)
}
\arguments{
\item{many_pops}{The return value from sim_proc_many_pops}

\item{verbose}{If TRUE, print details of failures}
}
\description{
Inspects the return value from sim_proc_many_pops to check for any
deviations from the specification.
}
\details{
There are no direct tests of this function based on manually created
datasets. The reason is that there are so many steps and most of those
steps already have their own checks. This this function is just designed to
check for obvious major deviations.
}
