% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsum.R
\name{dcollection_to_df}
\alias{dcollection_to_df}
\title{Convert dcollection into a data.frame}
\usage{
dcollection_to_df(dcollection)
}
\arguments{
\item{dcollection}{The dcollection to convert}
}
\description{
Convert dcollection into a data.frame
}
