% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_evaluator_homology.R
\name{fitness_evaluator_homology_fun}
\alias{fitness_evaluator_homology_fun}
\title{The homology based fitness_evaluator}
\usage{
fitness_evaluator_homology_fun(the_seq, comparators,
  h2fs = "h2fs_univariate_linear_fun")
}
\arguments{
\item{the_seq}{All the sequences to which fitnesses must be assigned.}

\item{comparators}{The sequences to which the input sequences must be compared}

\item{h2fs}{A function that remaps the homology results between the input
sequences and the comparators to a fitness score.}
}
\value{
A list with a single element called fitness.
}
\description{
Assigns each sequences a fitness based on the homology to provided sequences
}
