% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_evaluator_uniform.R
\name{fitness_evaluator_uniform}
\alias{fitness_evaluator_uniform}
\title{The uniform fitness_evaluator}
\usage{
fitness_evaluator_uniform(the_seq)
}
\arguments{
\item{the_seq}{All the sequences to which fitnesses must be assigned.}
}
\value{
A list with a single element called fitness.
}
\description{
Assigns each sequences a fitness uniformly sampled from (0,1)
}
