% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_processing_metrics.R
\name{fitness_processing_metrics_to_df}
\alias{fitness_processing_metrics_to_df}
\title{Converts a fitness_processing_metrics list to a data.frame}
\usage{
fitness_processing_metrics_to_df(fitness_processing_metrics)
}
\arguments{
\item{fitness_processing_metrics}{The list containing the fitness
processing metrics as produced by \code{sim_proc_many_pops}}
}
\description{
Converts a fitness_processing_metrics list to a data.frame
}
