% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\name{gfo_internal_Rdf}
\alias{gfo_internal_Rdf}
\title{Internal function implementing get_fit_offspring using data.frames}
\usage{
gfo_internal_Rdf(genealogy, fitness_score)
}
\description{
Internal function implementing get_fit_offspring using data.frames
}
