% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\name{gfo_internal_Rvec}
\alias{gfo_internal_Rvec}
\title{Internal function implementing get_fit_offspring using vectors}
\usage{
gfo_internal_Rvec(genealogy, fitness_score)
}
\description{
Internal function implementing get_fit_offspring using vectors
}
