% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_evaluator_homology.R
\name{h2fs_univariate_linear_fun}
\alias{h2fs_univariate_linear_fun}
\title{Univariate linear conversion between homology and fitness_score}
\usage{
h2fs_univariate_linear_fun(dists)
}
\arguments{
\item{dists}{The vector (of length one in this case) of distances between the sequence and the comparators.}
}
\value{
A single numeric value that is the fitness score
}
\description{
Just returns the input value since the homology values are scaled to the unit interval.
}
