% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genealogy.R
\name{make_genealogy}
\alias{make_genealogy}
\title{Constructs a genealogy}
\usage{
make_genealogy(ancestors)
}
\arguments{
\item{ancestors}{Character vector containing the sequences of the ancestors of the genealogy.}
}
\description{
Initializes a new genealogy from the ancestor sequences.
}
