% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_evaluator_homology.R
\name{max_homology}
\alias{max_homology}
\title{Computes the max homology possible to a set of comparator sequences.}
\usage{
max_homology(comparators)
}
\arguments{
\item{comparators}{A vector of character strings that form the comparators. Note the special restrictions placed on the usage of X.}
}
\value{
A list with an element for each comparator. This element contains the max possible homology between a sequence and the comparator.
}
\description{
Given comparator sequences, compute the max homology that can be achieved to these comparators given the conventions around the handling of X characters. See github issue #43
}
