% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoiseCache.R
\name{memoiseCache}
\alias{memoiseCache}
\title{Cache a function call with an associated seed value}
\usage{
memoiseCache(fun, args, cacheNamePrefix = NULL, seed = NULL, ...)
}
\arguments{
\item{fun}{The name of the function to execute as a character string}

\item{args}{A list of the arguments used by do.call to run fun.}

\item{cacheNamePrefix}{Optional prefix that gets added to the cacheName. Frequent usage is a diagnostic marker for OCD.}

\item{seed}{A positive integer that allows extra control over whether or not to compute or load the value from a cache}
}
\description{
Uses simpleCache::simpleCache to implement a form of memoisation. It caches a function's result based on the arguments provided and will reload that value instead of recomputing the function if the arguments match. Additionally allows the user to specify a seed value so that different results based on the same arguments.
}
