% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutator_general.R
\name{mutator_checks_general}
\alias{mutator_checks_general}
\title{General mutation checker}
\usage{
mutator_checks_general(fun, args)
}
\arguments{
\item{fun}{The mutator function}

\item{args}{A list of arguments with which the mutator such run 'normally'. i.e. a per base mutation rate between 0% and 5%. This will be used to call the mutator on dummy data during the checks.}
}
\description{
Performs basic sanity checks on a mutator and its candidate argument set.
}
\details{
Since mutators are custom functions that are only used in deeply nested code
and called with mechanisms like do.call, it makes debugging them really
hard. This function performs a number of checks to ensure that the basic
errors are flagged before the main loop is initiated.
}
