% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutator_uniform.R
\name{mutator_uniform_fun}
\alias{mutator_uniform_fun}
\title{The uniform mutator}
\usage{
mutator_uniform_fun(parent, mu)
}
\arguments{
\item{mu}{A single number between 0 and 1 specifying the per base mutation rate}

\item{patent}{A character string containing the DNA sequence of the parent}
}
\description{
Mutates a sequence giving each base an equal chance to experience a random mutation
}
