% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombination.R
\name{recombine_seqs}
\alias{recombine_seqs}
\title{Recombines two sequences}
\usage{
recombine_seqs(target_seq, recombination_partner_seq,
  min_dist_to_edge = 5)
}
\arguments{
\item{target_seq}{The sequence that will be replaced by the recombination}

\item{recombination_partner_seq}{The recombination partner sequence}
}
\description{
Recombines two sequences
}
