% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pop.R
\name{sim_pop}
\alias{sim_pop}
\title{Simulates evolution of DNA of a population}
\usage{
sim_pop(ancestors, n_gen = NULL, n_pop = NULL)
}
\arguments{
\item{ancestors}{A list of DNA sequences with which to start the population. Include the same sequence multiple times to achieve a target ratio.}
}
\description{
Produce a set of DNA sequences by simulation a population with multiple generations allowing mutation and recombination
}
\details{
Details: Simulated multiple generations
}
