% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_dmat.R
\name{summarize_dmat}
\alias{summarize_dmat}
\title{Summarize distance matrix}
\usage{
summarize_dmat(dmat)
}
\arguments{
\item{dmat}{A distance matrix, or any set of observations from a
distribution really.}
}
\description{
Summarizes a distance matrix into an average pairwise distance (`avg_hd`),
a set of percentile (`perc`) and a density estimate (`dens`). No
identifiers are added, that is not part of this function.
}
