% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genealogy.R
\name{check_genealogy_gen_num}
\alias{check_genealogy_gen_num}
\title{Check the gen_num column in a genealogy}
\usage{
check_genealogy_gen_num(genealogy, result = list())
}
\arguments{
\item{genealogy}{The genealogy to check.}

\item{result}{The list to which the result will be added and from which
previous result will be drawn to check the prerequisites.}
}
\value{
A list with TRUE or FALSE indicating whether the related check
passed.
}
\description{
Checks that gen_num is not missing, is a natural number and contains all
numbers between zero and the max gen_num.
}
