% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage.R
\name{linkage_diseq}
\alias{linkage_diseq}
\title{Computes the linkage disequilibrium for a set of sequences}
\usage{
linkage_diseq(seqs, min_prev = 0.2, max_prev = 0.8, verbose = FALSE)
}
\arguments{
\item{seqs}{The sequences on which the linkage disequilibrium is to be
computed}

\item{min_prev}{The minimum prevalence the majority variant must attain to
be considered for inclusion. (This cutoff will not see much action)}

\item{max_prev}{Majoritity variants with prevalences higher than this will
not be considered.}

\item{verbose}{Should intermediate results be included in the return value?}
}
\value{
A list with multiple elements:
\itemize{
  \item linkage The linkage scores for various distances and starting
  positions.
  \item cm The consensusMatrix. Only included if verbose is specified.
}
}
\description{
Computes the linkage disequilibrium for a set of sequences
}
