% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genealogy.R
\name{make_genealogy}
\alias{make_genealogy}
\title{Constructs a genealogy}
\usage{
make_genealogy(ancestors)
}
\arguments{
\item{ancestors}{If no information other than the initial ancestors are available, just pass in the character vector with the ancestors and a genealogy will be intialized.}
}
\description{
Makes a genealogy data structure and (not yet) checks that it is correct
}
