% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_next_gen.R
\name{sim_next_gen}
\alias{sim_next_gen}
\title{Simulate the next generation}
\usage{
sim_next_gen(parents, gen_size)
}
\arguments{
\item{parents}{A character vector with each element being a parent sequence}

\item{gen_size}{The size of each generation. Currently only allowed to be a single integer. Default value is 2.}
}
\description{
Simulates the next generation by calling a mutator on each parent
}
