% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_next_gen.R
\name{sim_next_gen}
\alias{sim_next_gen}
\title{Simulate the next generation}
\usage{
sim_next_gen(parents, gen_size, mutator)
}
\arguments{
\item{parents}{A character vector with each element being a parent sequence}

\item{gen_size}{The size of each generation. Currently only allowed to be a
single integer. Default value is 2.}

\item{mutator}{A list with two elements fun and args specifying the function that mutates parents into their offspring and the list of arguments said function requires.}
}
\value{
Currently, a list with a single element which is a character vector with each element being the sequence of a single offspring. In future, this should be expanded to include much more.
}
\description{
Simulates the next generation by calling a mutator on each parent
}
