% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_next_gen.R
\name{sim_next_gen}
\alias{sim_next_gen}
\title{Simulate the next generation}
\usage{
sim_next_gen(genealogy, gen_size, mutator, gen_num = 1)
}
\arguments{
\item{genealogy}{A genealogy data structure holding all the previous generations. TODO: link to general documentation. GH issue #6.}

\item{gen_size}{The size of each generation. Currently only allowed to be a
single integer. Default value is 2.}

\item{mutator}{A list with two elements fun and args specifying the function that mutates parents into their offspring and the list of arguments said function requires.}

\item{gen_num}{The generation number. Defaults to 1.}
}
\value{
A genealogy data structure. TODO: link to general documentation on a genealogy. GH issue #6
}
\description{
Simulates the next generation by calling a mutator on each parent
}
