% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pop.R
\name{sim_pop}
\alias{sim_pop}
\title{Simulates evolution of DNA of a population}
\usage{
sim_pop(ancestors, gen_size = 2, n_gen = NULL, n_pop = NULL)
}
\arguments{
\item{ancestors}{A list of DNA sequences with which to start the population. Include the same sequence multiple times to achieve a target ratio.}

\item{gen_size}{The size of each generation. Currently only allowed to be a single integer. Default value is 2.}

\item{n_gen}{The number of generations to simulate.}

\item{n_pop}{Stop the simulation when the population size exceeds this number.}
}
\description{
Produce a set of DNA sequences by simulation a population with multiple generations allowing mutation and recombination
}
\details{
Details: Simulated multiple generations
}
