% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_proc_many_pops.R
\name{sim_proc_many_pops}
\alias{sim_proc_many_pops}
\title{Simulate and process many populations}
\usage{
sim_proc_many_pops(arg_collection)
}
\arguments{
\item{arg_collection}{The collection of argument sets to be used in the
simulations}
}
\description{
Simulate and process many populations
}
