% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_proc_many_pops.R
\name{sim_proc_many_pops}
\alias{sim_proc_many_pops}
\title{Simulate and process many populations}
\usage{
sim_proc_many_pops(arg_collection, n_sims = 1, output_dmat = FALSE)
}
\arguments{
\item{arg_collection}{The collection of argument sets to be used in the
simulations}

\item{n_sims}{The number of times sim_pop must be called for each arg_set.
The total number of data sets produced will be \code{n_sims *
length(arg_collection)} if no fitness processing is specified and \code{n_sims *
length(arg_collection) * 2} if any fitness processing is specified.}

\item{output_dmat}{If TRUE, the distance matrices will be included in the
output. Uses lots of memory - default is FALSE.}
}
\value{
A list with the following elements:
\itemize{
  \item \code{dcollection}: A \code{dcollection} as checked by \code{\link{check_dcollection}}
  \item \code{arg_collection}: The input \code{arg_collection} as checked by
  \code{\link{check_arg_collection}}.
  \item \code{all_genealogies} [OPTIONAL]: A \code{data.frame} that contains all the
  genealogies, or in the case that \code{last_gen_only} was specified, only the
  last generations of the genealogies and all the identifier columns.
  \item \code{all_dmats} [OPTIONAL]: A list of lists containing the
  distance matrixes and their indentifiers.
  \item \code{n_sims}: The input \code{n_sims} indicating the number of
  datasets you can expect.
}
}
\description{
Simulate and process many populations
}
