% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsum.R
\name{summarize_dmat}
\alias{summarize_dmat}
\title{Summarize distance matrix}
\usage{
summarize_dmat(dmat)
}
\arguments{
\item{dmat}{A distance matrix, or any set of observations from a
distribution really.  @export}
}
\description{
Summarizes a distance matrix into an average pairwise distance
(\code{avg_hd}), a set of percentile (\code{perc}) and a density estimate
(\code{dens}). No identifiers are added, that is not part of this function.
}
\details{
Additionally, the distance matrix is divided into two clusters using the
\code{clara} function. Using this clustering, the within and between cluster
distances are computed and added into the final result.
}
