% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsum.R
\name{summarize_dmat}
\alias{summarize_dmat}
\title{Summarize distance matrix}
\usage{
summarize_dmat(dmat)
}
\arguments{
\item{dmat}{A distance
matrix, or any set of observations from a distribution really.  @export}
}
\value{
A list with the following elements: 
\itemize{ 
  \item avg_hd The average of the pairwise distances in the distance matrix.
  \item sd_hd The standard deviation of the pairwise distances in the distance matrix.
  \item perc The percentiles of the distribution of the pairwise distances.
  \item dens The result obtained when calling \code{density} on the distribution.
  \item clara2 The summary of the metrics related to the clustering of the distance matrix.
}
}
\description{
Summarizes a distance matrix into a basic set of metrics. The full list of
metrics is described in the section detailing the return value of this
function. Additionally, the distance matrix is divided into two clusters
using the \code{clara} function. Using this clustering, the within and
between cluster distances are computed and added into the final result.
}
